requirements.txt
allure-pytest>=2.8.40
allure-python-commons>=2.8.40
configparser>=5.0.2
configreader>=0.0.6
openpyxl>=3.0.9
packaging>=20.9
Pillow>=8.2.0
PyAutoGUI>=0.9.53
pytest>=6.2.3
pytest-ast-transformer>=1.0.3
pytest-forked>=1.3.0
pytest-html>=3.1.1
pytest-metadata>=1.11.0
pytest-soft-assertions>=0.1.2
pytest-xdist>=2.2.1
requests>=2.25.1
selenium==3.141.0
seletools>=1.2.1
urllib3>=1.26.4
virtualenv>=20.4.7
virtualenv-clone>=0.5.7
webdriver-manager>=3.3.0

